refactor(winglet,canard): close FCA boundary violations introduced on this branch
Cross-review (PR #328) confirmed 106 FCA findings; 10 of them were introduced by
this branch. This commit closes 9 and records the decision for the 10th.

Bench directories (FIX-097/098/099/100)
- The 16 new bench files reached into `src` internals while the same branch was
  narrowing those entry points to named re-exports. Routed every bench import
  through the package entry point instead, so the benchmark measures the surface
  consumers actually get.
- `Operation` is now a named export of `patch/index.ts`; the json bench could not
  build a patch through the entry point without it.
- `bench` is declared outside FCA management via `structure.additionalExcludedDirectories`.
  It is measurement tooling, and no fractal ancestor exists to own the shared
  `driveRenders` helper.

Internal helpers moved into organs (FIX-091/092/093)
- `countDecimals`, `countRetainedKeys`, `equalsBuiltin`, `differenceRecursive` are
  absent from their fractals' public exports, so they belong one level down.
- `isForwardRefComponent` / `isLazyComponent` deliberately stay put: they are public
  alongside four sibling predicates, and moving two of six would split the directory.
  Tracked with the rest of the zero-peer-file cleanup.

Patch vocabulary extracted to a sibling fractal (FIX-086)
- `patch/index.ts` re-exported its children while those children imported
  `patch/type.ts`, closing a `patch -> applyPatch -> patch` cycle. Moved the shared
  vocabulary to `JSONPointer/utils/patchModel`, split by concern into `constant.ts`
  (Operation) and `type.ts` (patch shapes), with INTENT.md and DETAIL.md.
- Relative specifiers are used on purpose: the adapter drops `@/json/*` alias edges
  from the dependency graph, so an alias would hide the crossing rather than fix it.
- Resolves the cycle plus six pre-existing external-import-boundary errors.

Acceptance criteria (FIX-017)
- `intersectSchema/DETAIL.md` gained the missing section, recording the
  `multipleOf` non-finite contract this branch changed.

Verified: typecheck 0 for common-utils, json, react-utils, schema-form; tests
1079/1079, 551/551, 183/183. `structure_validate` on `json/src` reports zero
circular-dependency and zero external-import-boundary errors.

Remaining findings tracked in #329, #330, #331.
