+++ .filid/config.json
@@ -3,9 +3,7 @@
   "language": "ko",
   "adapters": {
     "mode": "auto",
-    "enabled": [
-      "ecmascript"
-    ]
+    "enabled": ["ecmascript"]
   },
   "rules": {
     "organ-no-intentmd": {
@@ -15,9 +13,7 @@
     "module-entry-point": {
       "enabled": true,
       "severity": "warning",
-      "exempt": [
-        "**/canard/schema-form/src/app"
-      ]
+      "exempt": ["**/canard/schema-form/src/app"]
     },
     "max-depth": {
       "enabled": true,
@@ -34,13 +30,11 @@
     "zero-peer-file": {
       "enabled": true,
       "severity": "warning",
-      "exempt": [
-        "**/canard/schema-form/**",
-        "**/lerx/promise-modal/**"
-      ]
+      "exempt": ["**/canard/schema-form/**", "**/lerx/promise-modal/**"]
     }
   },
   "structure": {
+    "additionalExcludedDirectories": ["bench"],
     "additionalAllowedPeers": [
       {
         "basename": "type.ts"
