+++ packages/winglet/common-utils/src/utils/math/lcm.ts
@@ -1,6 +1,6 @@
 import { abs } from './abs';
 import { MAX_FRACTION_DIGITS } from './constant';
-import { countDecimals } from './countDecimals';
+import { countDecimals } from './utils/countDecimals';
 import { gcd } from './gcd';
 import { maxLite } from './maxLite';
