+++ packages/winglet/common-utils/src/utils/object/equals.ts
@@ -2,8 +2,8 @@ import { OBJECT_TAG } from '@/common-utils/constant/typeTag';
 import { getTypeTag } from '@/common-utils/libs/getTypeTag';
 import { hasOwnProperty } from '@/common-utils/libs/hasOwnProperty';
 
-import { countRetainedKeys } from './countRetainedKeys';
-import { equalsBuiltin } from './equalsBuiltin';
+import { countRetainedKeys } from './utils/countRetainedKeys';
+import { equalsBuiltin } from './utils/equalsBuiltin';
 
 /**
  * Sentinel used to bypass internal-state tag checks for current-realm
