+++ packages/winglet/json/src/JSONPointer/utils/patch/applyPatch/__tests__/applyPatch.copyOnWrite.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
-import type { Patch } from '../../type';
-import { Operation } from '../../type';
+import type { Patch } from '../../../patchModel';
+import { Operation } from '../../../patchModel';
 import { applyPatch } from '../applyPatch';
 
 describe('applyPatch copy-on-write', () => {
