+++ packages/winglet/json/src/JSONPointer/utils/patch/applyPatch/__tests__/applyPatch.pathForms.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { Operation, type Patch } from '../../type';
+import { Operation, type Patch } from '../../../patchModel';
 import { applyPatch } from '../applyPatch';
 import { JsonPatchError } from '../utils/error';
