+++ packages/winglet/json/src/JSONPointer/utils/patch/applyPatch/utils/handleArray.ts
@@ -4,7 +4,7 @@ import { getValue } from '@/json/JSONPointer/utils/manipulator/getValue';
 import { setValue } from '@/json/JSONPointer/utils/manipulator/setValue';
 import type { JsonArray, JsonRoot } from '@/json/type';
 
-import { Operation, type Patch } from '../../type';
+import { Operation, type Patch } from '../../../patchModel';
 import { JsonPatchError } from './error';
 import { isCircularMoveReference } from './isCircularMoveReference';
