+++ packages/winglet/json/src/JSONPointer/utils/patchModel/index.ts
@@ -1,0 +1,10 @@
+export { Operation } from './constant';
+export type {
+  Patch,
+  TestPatch,
+  AddPatch,
+  ReplacePatch,
+  RemovePatch,
+  CopyPatch,
+  MovePatch,
+} from './type';
